cloud/functions/chat_history/pyproject.toml
[tool.poetry]
name = "chat-history"
version = "0.1.0"
description = ""
authors = ["Andrey Kuznetsov <andrey_kuznetsov@epam.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.0.1"
flask_cors = "^5.0.1"
google-cloud-storage = "^2.1.0"
google-auth = "^2.28.1"
google-cloud-secret-manager = "^2.23.3"
requests = "^2.31.0"
firebase-admin = "^6.9.0"
colab-utils = { version = "^0.0.11", source = "gcp" }
#solvia-utils = {path = "../solvia_utils", develop = true}

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "gcp"
url = "https://europe-southwest1-python.pkg.dev/colab-invest-helper/simple/simple/"
priority = "supplemental"
